regen as per fa82bf13

+++ src/operators/Box/PyBoxAttributes.C
@@ -116,7 +116,7 @@ BoxAttributes_SetAmount(PyObject *self, PyObject *args)
     long val = PyLong_AsLong(args);
     int cval = int(val);
 
-    if ((val == -1 && PyErr_Occurred()) || cval != val)
+    if ((val == -1 && PyErr_Occurred()) || long(cval) != val)
     {
         Py_XDECREF(packaged_args);
         PyErr_Clear();
