+++ src/operators/MetricThreshold/PyMetricThresholdAttributes.C
@@ -233,7 +233,7 @@ MetricThresholdAttributes_SetPreset(PyObject *self, PyObject *args)
     long val = PyLong_AsLong(args);
     int cval = int(val);
 
-    if ((val == -1 && PyErr_Occurred()) || cval != val)
+    if ((val == -1 && PyErr_Occurred()) || long(cval) != val)
     {
         Py_XDECREF(packaged_args);
         PyErr_Clear();
