+++ src/operators/ToroidalPoloidalProjection/PyToroidalPoloidalProjection.C
@@ -244,7 +244,7 @@ ToroidalPoloidalProjection_SetCentroidSource(PyObject *self, PyObject *args)
     long val = PyLong_AsLong(args);
     int cval = int(val);
 
-    if ((val == -1 && PyErr_Occurred()) || cval != val)
+    if ((val == -1 && PyErr_Occurred()) || long(cval) != val)
     {
         Py_XDECREF(packaged_args);
         PyErr_Clear();
